requirements.txt
aniso8601==8.0.0
attrs==19.3.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-restplus==0.13.0
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-speech==1.2.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.0.0
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyrsistent==0.15.6
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
zipp==0.6.0
